requirements.txt
transformers
soundfile
torchvggish
mlflow
torchsummary